Modify implementation of Signals
